Moved helper functions to utils.js
