+++ lib/computeGeometry.js
@@ -1,10 +1,4 @@
-const PLACEMENT_OPTIONS = {
-  TOP: 'top',
-  RIGHT: 'right',
-  BOTTOM: 'bottom',
-  LEFT: 'left',
-  AUTO: 'auto',
-};
+import { PLACEMENT_OPTIONS, getArrowStyle } from './utils';
 
 function Point(x, y) {
   this.x = x;
@@ -45,46 +39,6 @@ const computeGeometry = ({
     default:
       return computeAutoGeometry(options);
   }
-};
-
-const getArrowStyle = ({
-  anchorPoint,
-  popoverOrigin,
-  arrowSize,
-  placement,
-}) => {
-  let width;
-  let height;
-  // swap width and height for left/right positioned tooltips
-  if (
-    placement === PLACEMENT_OPTIONS.TOP ||
-    placement === PLACEMENT_OPTIONS.BOTTOM
-  ) {
-    width = arrowSize.width;
-    height = arrowSize.height * 2;
-  } else {
-    width = arrowSize.height * 2;
-    height = arrowSize.width;
-  }
-
-  return {
-    width,
-    height,
-    left: anchorPoint.x - popoverOrigin.x - width / 2,
-    top: anchorPoint.y - popoverOrigin.y - height / 2,
-    borderTopWidth: height / 2,
-    borderRightWidth: width / 2,
-    borderBottomWidth: height / 2,
-    borderLeftWidth: width / 2,
-    borderTopColor:
-      placement === PLACEMENT_OPTIONS.TOP ? undefined : 'transparent',
-    borderRightColor:
-      placement === PLACEMENT_OPTIONS.RIGHT ? undefined : 'transparent',
-    borderBottomColor:
-      placement === PLACEMENT_OPTIONS.BOTTOM ? undefined : 'transparent',
-    borderLeftColor:
-      placement === PLACEMENT_OPTIONS.LEFT ? undefined : 'transparent',
-  };
 };
 
 const computeTopGeometry = ({
